(chore): rename the package for camel-test-infra-kafka for greater consistency
